refactor(storage): extract 5 no-gate stores from db.rs (brief, audit, settings, egress, folders)
Pure MOVEs, zero behavior change (the storage/mcp_store.rs pattern):
db.rs -1365 lines (24,905 -> 23,540), 5 new storage/*_store.rs
(inherent impl Db split), storage/mod.rs +5 registrations.

- brief_store.rs: brief_schedules + brief_runs CRUD + migrate_briefs +
  row mappers (the seal-coupled purge_pending_brief_runs_tx LEFT in db.rs).
- audit_store.rs: audit-findings + schedule CRUD + migrate_audit + the
  pending-purge _tx helpers (pub(crate), still called from the 7 db.rs
  seal sites cross-file). Raw-row store; the visibility re-gate lives in
  the command/audit layer, not here.
- settings_store.rs: the kv get/set/all_setting.
- egress_store.rs: the content-free egress-ledger methods.
- folders_store.rs: folder/note-folder CRUD + tree + schema (33 methods);
  the gated/seal-coupled folder readers (documents_in_folder,
  blank_sealed_notes_in_folders, etc.) LEFT in db.rs.

No lock surface: verified per-method that no moved method calls
visibility_clause / meeting_is_unlocked / seal-encrypt content (the seal
purges + gated readers stay in db.rs). add_column_if_missing promoted
pub(crate) for migrate_audit; map_err/lock already were. Tests stay in
db.rs (count 250 = trunk). Verified: cargo test --lib brief/audit/egress/
setting/folder + full storage::db 250/0, clippy --lib -D warnings clean,
MSRV grep clean.
