+++ src-tauri/src/storage/mod.rs
@@ -1,8 +1,13 @@
+pub(crate) mod audit_store;
+pub(crate) mod brief_store;
 pub mod db;
+pub(crate) mod egress_store;
+pub(crate) mod folders_store;
 pub(crate) mod links;
 pub(crate) mod mcp_store;
 pub mod migration;
 pub mod models;
+pub(crate) mod settings_store;
 pub mod usage;
 
 pub use db::*;
